Put a bit of order in the src directory, group and name things appropriately
