+++ platformio.ini
@@ -65,7 +65,7 @@ lib_deps =
 
 build_flags = ${env.build_flags} -Os 
 # -DRADIOLIB_GODMODE
-build_src_filter = ${env.build_src_filter} -<portduino/>
+build_src_filter = ${env.build_src_filter} -<platform/portduino/>
 
 ; Common libs for communicating over TCP/IP networks such as MQTT 
 [networking_base]
@@ -92,7 +92,7 @@ lib_deps =
 extends = arduino_base
 platform = espressif32@3.5.0
 build_src_filter = 
-  ${arduino_base.build_src_filter} -<nrf52/> -<stm32wl> -<rp2040>
+  ${arduino_base.build_src_filter} -<platform/nrf52/> -<platform/stm32wl> -<platform/rp2040>
 upload_speed = 115200
 debug_init_break = tbreak setup
 
@@ -101,7 +101,7 @@ debug_init_break = tbreak setup
 # This overrides the BLE logging default of LOG_LEVEL_INFO (1) from: .pio/libdeps/tbeam/NimBLE-Arduino/src/esp_nimble_cfg.h
 # -DUSE_NEW_ESP32_BLUETOOTH will enable the new NimBLE C++ api
 build_flags =
-  ${arduino_base.build_flags} -Wall -Wextra -Isrc/esp32 -Isrc/esp32-mfix-esp32-psram-cache-issue -lnimble -std=c++11 
+  ${arduino_base.build_flags} -Wall -Wextra -Isrc/platform/esp32 -lnimble -std=c++11 
   -DLOG_LOCAL_LEVEL=ESP_LOG_DEBUG -DCORE_DEBUG_LEVEL=ARDUHAL_LOG_LEVEL_DEBUG -DMYNEWT_VAL_BLE_HS_LOG_LVL=LOG_LEVEL_CRITICAL
   -DAXP_DEBUG_PORT=Serial -DUSE_NEW_ESP32_BLUETOOTH -DCONFIG_BT_NIMBLE_ENABLED -DCONFIG_NIMBLE_CPP_LOG_LEVEL=1
 lib_deps =
@@ -142,9 +142,9 @@ build_type = debug ; I'm debugging with ICE a lot now
 ; note: liboberon provides the AES256 implementation for NRF52 (though not using the hardware acceleration of the NRF52840 - FIXME)
 build_flags = 
   ${arduino_base.build_flags} -Wno-unused-variable 
-  -Isrc/nrf52
+  -Isrc/platform/nrf52
 build_src_filter = 
-  ${arduino_base.build_src_filter} -<esp32/> -<stm32wl> -<nimble/> -<mesh/wifi/> -<mesh/http/> -<modules/esp32> -<mqtt/> -<rp2040>
+  ${arduino_base.build_src_filter} -<platform/esp32/> -<platform/stm32wl> -<nimble/> -<mesh/wifi/> -<mesh/http/> -<modules/esp32> -<mqtt/> -<platform/rp2040>
 lib_ignore =
   BluetoothOTA
 
@@ -174,11 +174,11 @@ board_build.core = earlephilhower
 board_build.filesystem_size = 0.5m
 build_flags = 
   ${arduino_base.build_flags} -Wno-unused-variable 
-  -Isrc/rp2040
+  -Isrc/platform/rp2040
   -D__PLAT_RP2040__
 #  -D _POSIX_THREADS
 build_src_filter = 
-  ${arduino_base.build_src_filter} -<esp32/> -<nimble/> -<mesh/wifi/> -<mesh/http/> -<modules/esp32> -<mqtt/> -<nrf52/> -<stm32wl>
+  ${arduino_base.build_src_filter} -<platform/esp32/> -<nimble/> -<mesh/wifi/> -<mesh/http/> -<modules/esp32> -<mqtt/> -<platform/nrf52/> -<platform/stm32wl>
 lib_ignore =
   BluetoothOTA
 lib_deps =
