+++ src/platform/esp32/BluetoothSoftwareUpdate.cpp
