+++ src/platform/esp32/BluetoothSoftwareUpdate.h
