+++ src/platform/esp32/CallbackCharacteristic.h
