+++ src/platform/esp32/NimbleSoftwareUpdate.c
