+++ src/platform/esp32/SimpleAllocator.cpp
