+++ src/platform/esp32/SimpleAllocator.h
