+++ src/platform/esp32/architecture.h
