+++ src/platform/nrf52/aes-256/tiny-aes.h
