+++ src/platform/nrf52/alloc.cpp
