+++ src/platform/nrf52/architecture.h
