+++ src/platform/nrf52/hardfault.cpp
