+++ src/platform/nrf52/main-bare.cpp
