+++ src/platform/rp2040/architecture.h
