+++ variants/portduino/platformio.ini
@@ -2,11 +2,11 @@
 [portduino_base]
 build_src_filter = 
   ${env.build_src_filter} 
-  -<esp32/> 
+  -<platform/esp32/> 
   -<nimble/> 
-  -<nrf52/> 
-  -<stm32wl/> 
-  -<rp2040>
+  -<platform/nrf52/> 
+  -<platform/stm32wl/> 
+  -<platform/rp2040>
   -<mesh/http/> 
   -<modules/esp32>
   -<modules/Telemetry>
@@ -16,7 +16,7 @@ lib_deps =
   ${networking_base.lib_deps}
   rweather/Crypto
   https://github.com/meshtastic/RadioLib.git#5582ac30578ff3f53f20630a00b2a8a4b8f92c74
-build_flags = ${arduino_base.build_flags} -Isrc/portduino
+build_flags = ${arduino_base.build_flags} -Isrc/platform/portduino
 
 [env:native]
 platform = https://github.com/meshtastic/platform-native.git
